change made on line 1987
I corrected the English errors on line 1987 where the MC states "your perfect girlfriend, monika, will always love you!". 
I corrected this sentence and replaced it with proper English and a more fluent/ smooth sentence.
What I wrote: "As am I, your amazing Monika! I will always love you!"
I don't know how to do this too but I think it would only be fair if it displayed the player name when it said that as supposed to Monika. Unfortunately I don't know how to do this but it would only make sense for this to happen.